+++ R/deactivate.R
@@ -1,6 +1,8 @@
 #' @rdname activate
+#' @param clean If `TRUE`, will also remove the `renv/` directory and the
+#'   lockfile.
 #' @export
-deactivate <- function(project = NULL) {
+deactivate <- function(project = NULL, clean = FALSE) {
 
   renv_scope_error_handler()
 
@@ -10,6 +12,11 @@ deactivate <- function(project = NULL) {
   renv_infrastructure_remove_rprofile(project)
 
   unload(project)
+
+  if (clean) {
+    unlink(file.path(project, "renv.lock"))
+    unlink(file.path(project, "renv"), recursive = TRUE, force = TRUE)
+  }
 
   renv_restart_request(project, reason = "renv deactivated")
   invisible(project)